+++ daemon/labels_test.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/cilium/cilium/common"
 	"github.com/cilium/cilium/daemon/options"
+	"github.com/cilium/cilium/pkg/comparator"
 	"github.com/cilium/cilium/pkg/endpoint"
 	"github.com/cilium/cilium/pkg/kvstore"
 	"github.com/cilium/cilium/pkg/labels"
@@ -68,12 +69,11 @@ func (ds *DaemonSuite) SetUpTest(c *C) {
 	}
 	daemonConf.RunDir = tempRunDir
 	daemonConf.StateDir = tempRunDir
-	daemonConf.DockerEndpoint = "tcp://127.0.0.1"
 	// Get the default labels prefix filter
-	lblsPrefix, err := labels.ParseLabelPrefixCfg(nil, "")
+	err = labels.ParseLabelPrefixCfg(nil, "")
 	c.Assert(err, IsNil)
-	daemonConf.ValidLabelPrefixes = lblsPrefix
 	daemonConf.Opts.Set(endpoint.OptionDropNotify, true)
+	daemonConf.Opts.Set(endpoint.OptionTraceNotify, true)
 	daemonConf.Device = "undefined"
 
 	kvstore.SetupDummy()
@@ -88,7 +88,9 @@ func (ds *DaemonSuite) SetUpTest(c *C) {
 }
 
 func (ds *DaemonSuite) TearDownTest(c *C) {
-	os.RemoveAll(ds.d.conf.RunDir)
+	if ds.d != nil {
+		os.RemoveAll(ds.d.conf.RunDir)
+	}
 }
 
 func (ds *DaemonSuite) TestLabels(c *C) {
@@ -133,7 +135,7 @@ func (ds *DaemonSuite) TestLabels(c *C) {
 	wantSecCtxLbls.ID = policy.MinimalNumericIdentity
 	wantSecCtxLbls.Labels = lbls
 	c.Assert(gotSecCtxLbl.ID, Equals, wantSecCtxLbls.ID)
-	c.Assert(gotSecCtxLbl.Labels, DeepEquals, wantSecCtxLbls.Labels)
+	c.Assert(gotSecCtxLbl.Labels, comparator.DeepEquals, wantSecCtxLbls.Labels)
 	c.Assert(gotSecCtxLbl.RefCount(), Equals, 3)
 
 	err = ds.d.DeleteIdentity(policy.MinimalNumericIdentity, "containerLabel1-1")
@@ -143,7 +145,7 @@ func (ds *DaemonSuite) TestLabels(c *C) {
 	wantSecCtxLbls.ID = policy.MinimalNumericIdentity
 	wantSecCtxLbls.Labels = lbls
 	c.Assert(gotSecCtxLbl.ID, Equals, wantSecCtxLbls.ID)
-	c.Assert(gotSecCtxLbl.Labels, DeepEquals, wantSecCtxLbls.Labels)
+	c.Assert(gotSecCtxLbl.Labels, comparator.DeepEquals, wantSecCtxLbls.Labels)
 	c.Assert(gotSecCtxLbl.RefCount(), Equals, 2)
 
 	gotSecCtxLbl, err = ds.d.LookupIdentity(policy.MinimalNumericIdentity + 1)
@@ -151,7 +153,7 @@ func (ds *DaemonSuite) TestLabels(c *C) {
 	wantSecCtxLbls.ID = policy.MinimalNumericIdentity + 1
 	wantSecCtxLbls.Labels = lbls2
 	c.Assert(gotSecCtxLbl.ID, Equals, wantSecCtxLbls.ID)
-	c.Assert(gotSecCtxLbl.Labels, DeepEquals, wantSecCtxLbls.Labels)
+	c.Assert(gotSecCtxLbl.Labels, comparator.DeepEquals, wantSecCtxLbls.Labels)
 	c.Assert(gotSecCtxLbl.RefCount(), Equals, 2)
 
 	err = ds.d.DeleteIdentity(policy.MinimalNumericIdentity, "containerLabel1-2")
@@ -161,7 +163,7 @@ func (ds *DaemonSuite) TestLabels(c *C) {
 	wantSecCtxLbls.ID = policy.MinimalNumericIdentity
 	wantSecCtxLbls.Labels = lbls
 	c.Assert(gotSecCtxLbl.ID, Equals, wantSecCtxLbls.ID)
-	c.Assert(gotSecCtxLbl.Labels, DeepEquals, wantSecCtxLbls.Labels)
+	c.Assert(gotSecCtxLbl.Labels, comparator.DeepEquals, wantSecCtxLbls.Labels)
 	c.Assert(gotSecCtxLbl.RefCount(), Equals, 1)
 
 	err = ds.d.DeleteIdentity(policy.MinimalNumericIdentity, "containerLabel1-3")
@@ -175,7 +177,7 @@ func (ds *DaemonSuite) TestLabels(c *C) {
 	c.Assert(err, IsNil)
 
 	err = ds.d.DeleteIdentity(policy.MinimalNumericIdentity, "containerLabel1-non-existent")
-	c.Assert(err, DeepEquals, errors.New("identity not found"))
+	c.Assert(err, comparator.DeepEquals, errors.New("identity not found"))
 	gotSecCtxLbl, err = ds.d.LookupIdentity(policy.MinimalNumericIdentity)
 	c.Assert(err, IsNil)
 	c.Assert(gotSecCtxLbl, Equals, emptySecCtxLblPtr)
@@ -189,7 +191,7 @@ func (ds *DaemonSuite) TestLabels(c *C) {
 	sha256sum := lbls2.SHA256Sum()
 	gotSecCtxLbl, err = LookupIdentityBySHA256(sha256sum)
 	c.Assert(err, IsNil)
-	c.Assert(gotSecCtxLbl, DeepEquals, secCtxLbl)
+	c.Assert(gotSecCtxLbl, comparator.DeepEquals, secCtxLbl)
 
 	err = ds.d.DeleteIdentityBySHA256(sha256sum, "containerLabel2-1")
 	c.Assert(err, IsNil)
